refactor(types): move adapter types out of adapters/ directory
