+++ packages/next-safe-action/src/adapters.types.ts
@@ -1,24 +1,31 @@
+// Code courtesy of/highly inspired by https://github.com/decs/typeschema
+
 import type { GenericSchema, GenericSchemaAsync, InferInput, InferOutput } from "valibot";
+import type { InferType, Schema as YupSchema } from "yup";
 import type { z } from "zod";
 
 export type IfInstalled<T> = any extends T ? never : T;
-export type Schema = IfInstalled<z.ZodType> | IfInstalled<GenericSchema | GenericSchemaAsync>;
+
+export type Schema = IfInstalled<z.ZodType> | IfInstalled<GenericSchema | GenericSchemaAsync> | IfInstalled<YupSchema>;
+
 export type Infer<S extends Schema> =
 	S extends IfInstalled<z.ZodType>
 		? z.infer<S>
-		: S extends IfInstalled<GenericSchema>
+		: S extends IfInstalled<GenericSchema | GenericSchemaAsync>
 			? InferOutput<S>
-			: S extends IfInstalled<GenericSchemaAsync>
-				? InferOutput<S>
+			: S extends IfInstalled<YupSchema>
+				? InferType<S>
 				: never;
+
 export type InferIn<S extends Schema> =
 	S extends IfInstalled<z.ZodType>
 		? z.input<S>
-		: S extends IfInstalled<GenericSchema>
+		: S extends IfInstalled<GenericSchema | GenericSchemaAsync>
 			? InferInput<S>
-			: S extends IfInstalled<GenericSchemaAsync>
-				? InferInput<S>
+			: S extends IfInstalled<YupSchema>
+				? InferType<S>
 				: never;
+
 export type InferArray<BAS extends readonly Schema[]> = {
 	[K in keyof BAS]: Infer<BAS[K]>;
 };
@@ -44,6 +51,11 @@ export interface ValidationAdapter {
 	): Promise<{ success: true; data: Infer<S> } | { success: false; issues: ValidationIssue[] }>;
 	// valibot
 	validate<S extends IfInstalled<GenericSchema | GenericSchemaAsync>>(
+		schema: S,
+		data: unknown
+	): Promise<{ success: true; data: Infer<S> } | { success: false; issues: ValidationIssue[] }>;
+	// yup
+	validate<S extends IfInstalled<YupSchema>>(
 		schema: S,
 		data: unknown
 	): Promise<{ success: true; data: Infer<S> } | { success: false; issues: ValidationIssue[] }>;
